Add PyPI/CI/Ruff/Docs badges to README; fix quickstart float exponent
The quickstart's `10 ** df["exponent"]` raises when the exponent column
is negative int. Cast both sides to float.

+++ examples/quickstart.py
@@ -23,7 +23,9 @@ def main() -> None:
             formats=[],
         )
         # Convert raw mantissa to a human-readable price.
-        df["humanPrice"] = df["price"] * 10 ** df["exponent"]
+        # Cast to float — exponents are negative ints, and `int ** negative-int`
+        # raises in numpy/pandas. Float exponentiation handles it correctly.
+        df["humanPrice"] = df["price"].astype(float) * 10.0 ** df["exponent"].astype(float)
         print(df[["priceFeedId", "humanPrice", "publisherCount"]])
         print("update timestamp (µs):", df.attrs.get("timestampUs"))
